test set light

+++ dashboardController.cpp
@@ -50,10 +50,14 @@ void DashboardController::updateData(){
     setPm10Value(sensor->getValue(ODR_Interface::PM10));
     setNoiseValue(sensor->getValue(ODR_Interface::NOISE));
     setAtmValue(sensor->getValue(ODR_Interface::ATM));
-    uint32_t light_high = sensor->getValue(ODR_Interface::LUXH);
-    uint32_t light_low = sensor->getValue(ODR_Interface::LUXH);
+    uint16_t light_high = sensor->getValue(ODR_Interface::LUXH);
+    uint16_t light_low = sensor->getValue(ODR_Interface::LUXH);
+    // qDebug()<< "light high: " << light_high ;
+    // qDebug()<< "light low: " << light_low;
+    // qDebug()<< "light value : "<< (float) light_high << 16 | light_low;
 
-    setLightValue((float) (light_high<<16 | light_low));
+
+    setLightValue((float) light_high<<16 | light_low);
 
     QByteArray message = exportDataToJsonObject();
